requirements.txt
pillow
pre-commit
